add readable timespans
