+++ front/src/app/models/index.ts
@@ -3,3 +3,4 @@ export * from './ingredient.model';
 export * from './direction.model';
 export * from './page.model';
 export * from './login.model';
+export * from './timespan.model';
