+++ front/src/app/ui/index.ts
@@ -6,3 +6,4 @@ export * from './recipe-header';
 export * from './admin';
 export * from './common';
 export * from './login';
+export * from './pipes';
