+++ front/src/app/ui/pipes/index.ts
@@ -1,0 +1,1 @@
+export * from './timespan.pipe';
