+++ front/src/app/ui/recipe/recipe.component.ts
@@ -3,6 +3,7 @@ import { RecipesService } from '../../infra';
 import { IRecipe } from '../../models';
 import { Observable } from 'rxjs';
 import { ActivatedRoute } from '@angular/router';
+import { map } from 'rxjs/operators';
 
 @Component({
 	selector: 'app-recipe',
@@ -19,6 +20,11 @@ export class RecipeComponent implements OnInit {
 	) { }
 
 	public ngOnInit(): void {
-		this.recipe$ = this.recipesService.getRecipe(this.route.snapshot.paramMap.get('recipeId'));
+		this.recipe$ = this.recipesService.getRecipe(this.route.snapshot.paramMap.get('recipeId')).pipe(
+			map(r => {
+				r.preparationTime = '02:02:00'
+				return r;
+			})
+		);
 	}
 }
